Quantum Computing Project
To compile this program simply run the Makefile from the root directory of the project folder. Command: make This Makefile will create a bin folder and all of the compiled files will be placed there. After the makefile finishes the executable will be placed in the root of the bin folder. If you are in the root of the project folder you can run the program by typing: Command: ./bin/quantum